Added Code for ManyToMany Relationship

+++ src/main/java/com/batch12/rvirb/foundation/bank/service/AccountServiceImpl2.java
@@ -8,6 +8,7 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
 import com.batch12.rvirb.foundation.bank.entities.Account;
+import com.batch12.rvirb.foundation.bank.exceptions.AccountNotFound;
 import com.batch12.rvirb.foundation.bank.repositories.AccountRepository;
 
 //This Implementation will use CRUD Repository Layer, Actual Database
@@ -21,6 +22,10 @@ public class AccountServiceImpl2 implements AccountService {
 	
 	public Account getAccount(int id) {
 		Optional<Account> accountOpt = accountRepository.findById(id);
+		if (!accountOpt.isPresent()) {
+			throw new AccountNotFound("Id --" +id);
+		}
+		
 		Account account = accountOpt.get();
 		return account;
 		
@@ -39,6 +44,10 @@ public class AccountServiceImpl2 implements AccountService {
 
 	public Account updateAccount(int id, Account account) {
 		Optional<Account> accountOpt = accountRepository.findById(id);
+		if (!accountOpt.isPresent()) {
+			throw new AccountNotFound("Id --" +id);
+		}
+		
 		Account accountUpdated = accountOpt.get();
 		
 		accountUpdated.setAccountBalance(account.getAccountBalance());
@@ -51,6 +60,10 @@ public class AccountServiceImpl2 implements AccountService {
 
 	public Account deleteAccount(int id) {
 		Optional<Account> accountOpt = accountRepository.findById(id);
+		if (!accountOpt.isPresent()) {
+			throw new AccountNotFound("Id --" +id);
+		}
+		
 		Account accountDeleted = accountOpt.get();
 		
 		accountRepository.delete(accountDeleted);
